Fix running service lead to sticky notification
Stop service started using `BackgroundMode` before new video is played in foreground.
